pyproject.toml
[project]
name = "nats-semaphore"
version = "0.0.1"
description = "Distributed semaphore utility based on NATS JetStream key-value stores."
authors = [
    {name = "Mikołaj Nowak", email = "12396461+m3nowak@users.noreply.github.com"},
]
dependencies = ["nats-py<3,>=2.12", "async-timeout>=5.0.1"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ["I"]
fixable = ["ALL"]


[tool.pyright]
exclude = [
  "**/__pycache__",
]
include = ["src", "tests"]
pythonPlatform = "Linux"
pythonVersion = "3.10"
reportMissingImports = "error"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
format = "ruff format"
lint-fix = "ruff check --fix"
fix-all = {composite = ["lint-fix", "format"], keep_going = true}
types = "pyright src tests"

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "testcontainers[nats]>=4.13.3",
    "pytest-cov>=7.0.0",
]

